SelectWhichAddress  relocation
