+++ src/Include/Functions.php
@@ -205,65 +205,6 @@ function SelectWhichInfo($sPersonInfo, $sFamilyInfo, $bFormat = false)
     return $finalData;
 }
 
-//
-// Returns the correct address to use via the sReturnAddress arguments.
-// Function value returns 0 if no info was given, 1 if person info was used, and 2 if family info was used.
-// We do address lines 1 and 2 in together because seperately we might end up with half family address and half person address!
-//
-function SelectWhichAddress(&$sReturnAddress1, &$sReturnAddress2, $sPersonAddress1, $sPersonAddress2, $sFamilyAddress1, $sFamilyAddress2, $bFormat = false)
-{
-    if (SystemConfig::getValue('bShowFamilyData')) {
-        if ($bFormat) {
-            $sFamilyInfoBegin = "<span style='color: red;'>";
-            $sFamilyInfoEnd = '</span>';
-        }
-
-        if ($sPersonAddress1 || $sPersonAddress2) {
-            $sReturnAddress1 = $sPersonAddress1;
-            $sReturnAddress2 = $sPersonAddress2;
-
-            return 1;
-        } elseif ($sFamilyAddress1 || $sFamilyAddress2) {
-            if ($bFormat) {
-                if ($sFamilyAddress1) {
-                    $sReturnAddress1 = $sFamilyInfoBegin.$sFamilyAddress1.$sFamilyInfoEnd;
-                } else {
-                    $sReturnAddress1 = '';
-                }
-                if ($sFamilyAddress2) {
-                    $sReturnAddress2 = $sFamilyInfoBegin.$sFamilyAddress2.$sFamilyInfoEnd;
-                } else {
-                    $sReturnAddress2 = '';
-                }
-
-                return 2;
-            } else {
-                $sReturnAddress1 = $sFamilyAddress1;
-                $sReturnAddress2 = $sFamilyAddress2;
-
-                return 2;
-            }
-        } else {
-            $sReturnAddress1 = '';
-            $sReturnAddress2 = '';
-
-            return 0;
-        }
-    } else {
-        if ($sPersonAddress1 || $sPersonAddress2) {
-            $sReturnAddress1 = $sPersonAddress1;
-            $sReturnAddress2 = $sPersonAddress2;
-
-            return 1;
-        } else {
-            $sReturnAddress1 = '';
-            $sReturnAddress2 = '';
-
-            return 0;
-        }
-    }
-}
-
 function ChopLastCharacter($sText)
 {
     return mb_substr($sText, 0, strlen($sText) - 1);
