+++ src/PrintPastoralCare.php
@@ -114,7 +114,7 @@ $dBirthDate = OutputUtils::FormatBirthDate($per_BirthYear, $per_BirthMonth, $per
 
 // Assign the values locally, after selecting whether to display the family or person information
 
-SelectWhichAddress($sAddress1, $sAddress2, $per_Address1, $per_Address2, $fam_Address1, $fam_Address2, false);
+MiscUtils::SelectWhichAddress($sAddress1, $sAddress2, $per_Address1, $per_Address2, $fam_Address1, $fam_Address2, false);
 $sCity = SelectWhichInfo($per_City, $fam_City, false);
 $sState = SelectWhichInfo($per_State, $fam_State, false);
 $sZip = SelectWhichInfo($per_Zip, $fam_Zip, false);
