+++ src/Reports/GroupReport.php
@@ -89,7 +89,7 @@ if ($iMode == 1) {
 
         $pdf->sFamily = OutputUtils::FormatFullName($aRow['per_Title'], $aRow['per_FirstName'], $aRow['per_MiddleName'], $aRow['per_LastName'], $aRow['per_Suffix'], 3);
 
-        SelectWhichAddress($sAddress1, $sAddress2, $aRow['per_Address1'], $aRow['per_Address2'], $aRow['fam_Address1'], $aRow['fam_Address2'], false);
+        MiscUtils::SelectWhichAddress($sAddress1, $sAddress2, $aRow['per_Address1'], $aRow['per_Address2'], $aRow['fam_Address1'], $aRow['fam_Address2'], false);
 
         $sCity = SelectWhichInfo($aRow['per_City'], $aRow['fam_City'], false);
         $sState = SelectWhichInfo($aRow['per_State'], $aRow['fam_State'], false);
